requirements.txt
pandas==2.3.3
requests==2.31.0
python-dotenv==1.0.0